Split shared test code into separate files.
before:
71 tests, 439 assertions, 0 failures, 0 errors

after:
71 tests, 439 assertions, 0 failures, 0 errors
