+++ test/helper.rb
@@ -10,3 +10,5 @@ end
 require 'test/unit'
 require 'active_record'
 require "#{File.dirname(__FILE__)}/../init"
+
+require 'shared'
