+++ test/shared.rb
@@ -1,0 +1,7 @@
+# Common shared behaviour.
+module Shared
+  autoload :List,           'shared_list'
+  autoload :ListSub,        'shared_list_sub'
+  autoload :ZeroBased,      'shared_zero_based'
+  autoload :ArrayScopeList, 'shared_array_scope_list'
+end
